Bump phpstan (#1151)
* Bump PHPStan and related deps and fix some issues

- PHPstan to 2.x and related packages.
- Fix some of the newly reported issues.
- Update baseline.

* Support console 6.x

+++ lib/Color/Gradient.php
@@ -85,7 +85,7 @@ use function array_fill;
     {
         $percentile = $percentile < 0 ? 0 : min(100, abs($percentile));
 
-        $offset = round((count($this->colors) - 1) / 100 * $percentile);
+        $offset = (int)(round((count($this->colors) - 1) / 100 * $percentile));
 
         return $this->colors[$offset];
     }
